Allow NFS relations to have b values > 2^32; remove dead code
git-svn-id: https://svn.code.sf.net/p/msieve/code/trunk@1076 8c02911e-0617-4013-80d0-c3bcbcae04d5

+++ gnfs/sieve/sieve.h
@@ -28,7 +28,7 @@ extern "C" {
 
 /* dump one NFS relation to the savefile */
 
-void print_relation(savefile_t *savefile, int64 a, uint32 b, 
+void print_relation(savefile_t *savefile, int64 a, uint64 b, 
 		uint32 *factors_r, uint32 num_factors_r, 
 		uint32 large_prime_r[MAX_LARGE_PRIMES],
 		uint32 *factors_a, uint32 num_factors_a, 
@@ -40,7 +40,7 @@ uint32 fplog(uint32 k, double log_of_base);
 
 /* find the size of f(a,b) both in bits and in base log_base */
 
-int32 fplog_eval_poly(int64 a, uint32 b, mpz_t scratch,
+int32 fplog_eval_poly(int64 a, uint64 b, mpz_t scratch,
 			mpz_poly_t *f, double log_base,
 			uint32 *bits);
 
@@ -50,11 +50,11 @@ int32 fplog_eval_poly(int64 a, uint32 b, mpz_t scratch,
    of the dynamic range in one byte of the sieve array */
 
 double get_log_base(mpz_poly_t *poly, 
-			int64 a0, int64 a1, uint32 b);
+			int64 a0, int64 a1, uint64 b);
 
 uint32 read_last_line(msieve_obj *obj, mpz_t n);
 
-void write_last_line(msieve_obj *obj, mpz_t n, uint32 b);
+void write_last_line(msieve_obj *obj, mpz_t n, uint64 b);
 
 #ifdef __cplusplus
 }
